sequential-orchestration/requirements.txt
# Multi-Agent Sequential Orchestration with LangChain and Google Gemini
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=0.0.6
python-dotenv>=1.0.0
